Handle out-of-order completions in pingpong examples
Keep track of whether send and/or receive is pending in libibverbs
pingpong examples.  This avoids failures when the remote side receives
data and posts a send very quickly, and the local side completes the
receive before the previous send.  With the old code, this could
result in posting a send before the previous send completed, and
therefore overrun the send queue.

Signed-off-by: Roland Dreier <rolandd@cisco.com>