r4904: opensm fails to find HCA if port is down
If opensm is started with no arguments, the default algorithm
for finding a port to bind to will skip ports which are present
but the link is DOWN.  If there is only one port in the system,
no port is selected and opensm tries the default HCA name "mthca0"
which, if not present, confuses opensm and it exits.

The following patch changes the port selection so that the first
active port is selected, and if none, the first non-disabled port.

Signed-off-by: Ralph Campbell <ralphc@pathscale.com>
Signed-off-by: Hal Rosenstock <halr@voltaire.com>