r6674: If librdmacm has an error processing an ESTABLISHED event, it will report a CONNECT_ERROR event to the user.  However, the kernel will list the cm_id state as connected.  This can result in librdmacm receiving a DISCONNECTED event on a cm_id that it failed to fully connect from userspace.  This event should not be reported to the user.
This condition can occur when the remote side establishes a connection, then
disconnects before the local side processes the ESTABLISHED event.  When the
ESTABLISHED event is finally processed, the local side will get a failure
transitioning the QP, causing librdmacm to report CONNECT_ERROR.  Since the
connection was never reported as ESTABLISHED to the user, do not report the
DISCONNECTED event.

A similar situation exists between the CONNECT_RESPONSE and REJECTED events.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>