r8523: Add calls to the RDMA CM to return its list of RDMA devices.  The calls are similar to ibv_get_device_list() / ibv_free_device_list().
Without this patch, RDMA device contexts are handed to the user only after
they create an rdma_cm_id and bind it to a local device.  By exposing the
device list to the user, it makes it easier for the user to allocate device
specific resources (such as PDs, CQs, etc.) that are shared among multiple
rdma_cm_id's.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>