Add async_event callback function to struct ibv_context_ops
Add a callback function so low-level driver libraries can get a
callback with each async event retrieved by the user.  It allows the
bypass library to do WQ or CQ processing that needs to happen when a
fatal async event happens.  This async callback is similar to the
cq_event callback that already exists in libibverbs.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>
Signed-off-by: Roland Dreier <rolandd@cisco.com>