Update libmthca for new libibverbs driver handling
The latest libibverbs development tree changes how low-level drivers
are found.  The driver must be in a shared object like
libmthca-rdmav2.so in the ordinary library path, rather than
infiniband/mthca.so as for libibverbs 1.0.  In addition, the driver
must call ibv_register_driver() to pass its entry point to libibverbs,
rather than exporting an ibv_driver_init() function.

This patch adds autoconf tests to libmthca to detect whether it is
being built against a stable libibverbs 1.0 tree or the new libibverbs
development tree.  Then based on the result, it builds a library with
the appropriate name and with the correct driver initialization handling.

Signed-off-by: Roland Dreier <rolandd@cisco.com>