Use the chelsio device number to determine the type of adapter.
The existing code was using the sysfs files to obtain the adapter type
or hw revision.  This method wasn't portable to older distros because
it used sysfs entries not inserted by the rdma core, but rather by the
PCI subsystem.  Further, these files don't exist in RHEL4u4! This code
was implemented before it was known that the chelsio device numbers
actually encode the hw revision.

The simpler and portable solution is to just determine the hardware
revision based on the device number.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>