Add support to join IPOIB multicast groups
Add to the librdmacm an IPOIB port space that allows interoperability with
IPOIB multicast traffic.  Use of the RDMA_PS_IPOIB is limited to multicast
join/leave.

Rename the RDMA_UD_QKEY to RDMA_UDP_QKEY to signify that the qkey is only
used with the RDMA_PS_IPOIB port space.  Update mckey test program based on
patch from Or Gerlitz.

Signed-off-by: Or Gerlitz <ogerlitz@voltaire.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>