RDMA/cma: fix 32-bit user / 64-bit kernel mismatch issue
A 64-bit kernel will pad the size of the event structure to a
multiple of 8-bytes.  When using a 32-bit kernel, the structure is left
aligned to a 4-byte boundary.  This results in the userspace event structure
being too small because of the padding.  Fix this by increasing the padding
at the end of the userspace event structure.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>