Fix call to ibv_free_device_list() in pingpong examples
When a -d option to specify which device to use is passed to the
pingpong examples, they iterate through the device list by
incrementing the dev_list pointer.  This means that the call to
ibv_free_device_list() may not get the right pointer.

Fix this by using an index to iterate through the array and leaving
the dev_list pointer itself alone.

Signed-off-by: Roland Dreier <rolandd@cisco.com>