Initialize send queue entry ownership bits
We need to initialize the owner bit of send queue WQEs to hardware
ownership whenever the QP is modified from reset to init, not just
when the QP is first allocated.  This avoids having the hardware
process stale WQEs when the QP is moved to reset but not destroyed and
then modified to init again.

This is the same bug fixed in the kernel by Eli Cohen <eli@mellanox.co.il>.

Signed-off-by: Roland Dreier <rolandd@cisco.com>