Fix problem with inline WQE in post_send error flow
Suppose a consumer posts a list of two WQEs, with the second wqe in
the list being an INLINE which is too long.  In this case, post_send
jumps to "out" with: nreq = 1, inl positive, and size in the range
allowing blueflame. All the blueflame test conditions are met.
However, the cntl pointer now points to the invalid wqe, and this will
be "blueflamed".

Fix this by setting inl to 0 before jumping out of the loop.

Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Roland Dreier <rolandd@cisco.com>