libibumad: remove opened umad devices internal tracking.
Opened umad devices will be tracked by its file descriptor and not by
umad device number. This is simpler in general, doesn't formal API and
in this way multiple opening of the same device is supported. Also this
doesn't require internal device tracking in libibumad
(struct Port ports[] array). It is removed completely here, which turns
the library to be thread-safe.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>