IB/libipathverbs - Limit amount of memory ibv_cmd_post_send() needs
The gen2_basic test program attempts to post thousands of WRs
in one call. The ibv_cmd_post_send() function then tried to allocate
enough memory on the stack in order to write this to the kernel.
This patch limits the amount of memory needed by posting WRs in
smaller groups.

Signed-off-by: Ralph Campbell <ralph.campbell@qlogic.com>