Don't add an extra entry to CQs
With mlx4 hardware, there is no need to add an extra entry when
creating a CQ.  This potentially saves a lot of memory if a consumer
asks for an exact power of 2 entries.

This change works without changing the kernel mlx4_ib driver's ABI by
subtracting 1 from the number of CQ entries before passing the value
to the kernel; the kernel will add 1 and end up with the same value
actually used by libmlx4.

Based on work from  Jack Morgenstein <jackm@dev.mellanox.co.il>.

Signed-off-by: Roland Dreier <rolandd@cisco.com>