libipathverbs: zero context struct at allocation time (prep for additional context ops)
The ibv_context structure will be getting additional ops,
to be added at the end of the structure (and not as part of
the existing ibv_context_ops structure).

Reason: ibv_context_ops is declared directly as a member of ibv_context,
and not as a pointer.  Binaries compiled with previous libibverbs versions
will not be backwards compatible if we add new operations to ibv_context_ops,
since fields following the ops structure will move.

To enable adding new operations at the end of the existing ibv_context struct,
all driver libraries MUST zero their context structure at allocation time, so
that new ops will be NULL by default.

Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>