Convert hyphen to minus sign in ibv_query_pkey man page
A bare "-" in a man page will be rendered as a hyphen; to get a minus
sign, "\-" must be used.  Very pedantic people (or automatic checkers,
such as Debian's lintian tool) may notice the difference.  The man page
for ibv_query_pkey incorrectly wrote a negative return value as "-1".
Fix this to be the correct "\-1".

Signed-off-by: Roland Dreier <rolandd@cisco.com>