Rename ib_gid_t in mad.h to mad_gid_t to prevent name collision with ib_types.h
Ok, I added that as well, but that only reports a compile error when someone
uses it.  Without the USE_DEPRECATED_IB_GID_T macro "ib_gid_t" still conflicts
with ib_types.h.  I really want to fix that problem.  What I like about the
macro is if someone has a lot of uses in their code this gives them an easy way
to compile against the new version.

Signed-off-by: Ira K. Weiny <weiny2@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>