Update libmthca to handle new kernel ABI
The kernel mthca driver in 2.6.26 added support for userspace to ask
for a DMA attribute that causes in-flight DMA to be flushed when a
memory region is written to by the HCA.  This needs to be used for CQ
buffers to avoid memory ordering problems on some systems (SGI Altix
are the prime example).

Add the needed support in libmthca to use this feature by passing the
mr_attrs data to the kernel mthca driver when issuing reg_mr commands.

This also gets rid of annoying warnings like

    ib_mthca 0000:02:00.0: Process 'tgtd' did not pass in MR attrs.
    ib_mthca 0000:02:00.0:   Update libmthca to fix this.

Signed-off-by: Roland Dreier <rolandd@cisco.com>