libibcm: open correct ucm file
When opening the ucm* file, we need to select the correct file based
on the underlying libibverbs device.  We cannot base the file name
on the uverbs* file name, since iwarp devices do not result in ucm*
files.  Instead, we need to search for the correct file name based
on the device name.

This fixes a problem reported by Jeff Squires of Cisco running ib_cm
on a system with both iwarp and ib devices present.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>