add --loop_ports option to perfquery
Hey Sasha,

We have a switch here that does not report the AllPortSelect flag as a
capability.  It's pretty annoying typing each port on the switch or
always having to script around this one oddball switch we have.  So I
added an option --loop_ports for perfquery.  If you want to do something
to all the ports on the CA/Switch, but AllPortSelect isn't available, it
loops through all the available ports instead.

There was already a workaround in the tool for a CA that did not support
the AllPortSelect flag.  I get the feeling the workaround may have been
for a specific hardware, so I kept the workaround in there.

Al

--
Albert Chu
chu11@llnl.gov
Computer Scientist
High Performance Systems Division
Lawrence Livermore National Laboratory

From: Albert Chu <chu11@llnl.gov>
Date: Tue, 7 Oct 2008 14:05:54 -0700
Subject: [PATCH] support --loop_ports in perfquery

Signed-off-by: Albert Chu <chu11@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>