cmatose: avoid missing completions
cmatose uses a single CQ for send and receive completions.  It then
counts completions to determine if all sends and receives are done.
It's possible for a receive completion to be polled when the intent
is to count send completions.  (See server side polling for sends
done, then receives done.  The poll will get up to 8 completions,
which can lead to sends and receives being polled together.)

Fix this by separating the send and receive completions to their
own CQs to avoid any issues knowing what type of receives have been
polled from the CQ.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>