libibumad: get_ca() can call release_ca() with uninitialized data
In umad.c, get_ca(), the error path when the call to
scandir() fails calls release_ca(ca) when the structure
has not yet been initialized. This can cause free() to
be called on random memory addresses.
The fix is to initialize ca->numports and ca->ports
earlier in get_ca().

Signed-off-by: Ralph Campbell <ralph.campbell@qlogic.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>