infiniband-diags/src/ibnetdiscover.c missing LID information on --ports
In commit 496b83de4c4175ef3f8e4517708e1cbb8ee782f3 you added code to get the
correct information from the proper port, which fixes problems with wrong
speeds, etc. but appears to have broken the reporting for the LID in the
--ports output.
Without --ports, the values used for a switch are node->smalid
and node->smalmc but with --ports the value is taken from port->lid.  Also,
when getting the "other side" of a link, a quick look at the code seems like
it just uses port->lid as well.

I did a quick hack in get_node() to replace port->lid and port->lmc with the
values that were saved in node->smalid and node->smalmc and it seems to
solve the problem.  I just worry that I have not really thought through all
of the impacts of this.  Also, if my basic idea is correct, it seems like
other coding choices would be to either leave out lid/lmc decode
from decode_port_info() and do the decode in the two places where it is
needed, or to add a boolean argument to decode_port_info() to enable/disable
lid/lmc decode.

Dave

------- quick hack as a patch if you want it ------

    Fix problem where LID 0 is showing on switch ports

Signed-off-by: David McMillen <davem@systemfabricworks.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>