Create a new library libibnetdisc
>From e1c2c10678b0d1d90f7eb31eb1c1441b5ee43311 Mon Sep 17 00:00:00 2001
From: Ira Weiny <weiny2@llnl.gov>
Date: Fri, 3 Apr 2009 15:28:08 -0700
Subject: [PATCH] Create a new library libibnetdisc

This encompasses the functionality of ibnetdiscover in a C library.  It returns
a single "ibnd_fabric_t" object which represents the data found during the
scan.  The NodeInfo, PortInfo, and SwitchInfo are preserved from the queries
made on the fabric to be used by the calling function as they see fit.

This greatly benefits some diags like iblinkinfo.pl.  This diag in particular
was re-written using this library in C and has shown an 85% speed up on a ~1000
node cluster.

Previous iblinkinfo.pl
   real    3m35.876s
   user    0m13.210s
   sys     1m1.046s

New iblinkinfotest
   real    0m32.869s
   user    0m0.067s
   sys     0m0.140s

Signed-off-by: Ira Weiny <weiny2@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>