Fix ibidsverify.pl to use the correct cache file
Sasha,

I found this bug when I was testing the libibnetdisc stuff.

This applies to the master.

Ira

>From 656ad88a1f3ca6bcd7601b03da1b3822e4091156 Mon Sep 17 00:00:00 2001
From: Ira Weiny <weiny2@llnl.gov>
Date: Fri, 3 Apr 2009 16:00:46 -0700
Subject: [PATCH] Fix ibidsverify.pl to use the correct cache file

In addition add the -C and -P options for specifying a different HCA and port

Signed-off-by: Ira Weiny <weiny2@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>