ibnetdiscover: only report max hops discovered when requested
Added "-m" flag to report this information if the user wants it.  I also
changed the text in the message which says "reported max hops discovered".  I
don't know if we want to change that text to something else but I wanted to
indicate this number is not constant and may change.  This is true not just if
you change the algorithm of discovery but also if you run from different
nodes.

Thoughts,
Ira

From: Ira Weiny <weiny2@llnl.gov>
Date: Wed, 6 May 2009 17:56:23 -0700
Subject: [PATCH] ibnetdiscover: only report max hops discovered when requested

Signed-off-by: Ira Weiny <weiny2@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>