infiniband-diags: Do not change port state on SubnSet
Do not change port state on SubnSet

When changing physical state do not change logical port state to avoid invalid
transitions in port state machine.
From the IB spec When writing PortInfo:PortState, only legal transitions are
valid. So if PortState is ACTIVE and we try to set it to ACTIVE this will fail.

This patch will allow reset in a single MAD by changing physical state to polling.

Signed-off-by: Eli Dorfman <elid@voltaire.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>