infiniband_diags/ibportstate.c perform reset in a single mad
perform reset in a single mad

Perform reset by changing physical state to polling instead of
disable and enable.
This ensures that reset a port will not disconnect it (assuming negotiation succeeds).
Only disable operation is not allowed on HCA ports since it will irreversibly
disconnect it though this may also happen on a switch ingress port.

Signed-off-by: Eli Dorfman <elid@voltaire.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>