infiniband-diags/src/ibqueryerrors.c: Remove --all option and replace it with --switch, --ca, --router
By default ibqueryerrors should print errors for all node types.
Adding the other options allows for the limitation of this output.

Also change the --switch option to be --node-guid which is really more
accurate and use "-G" for better compliance with other utilities.  "-S"
is left in for backward compatibility for the time being.

Update the man page

Signed-off-by: Ira Weiny <weiny2@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>