libnes: Fix a thread safety problem on send and recv paths.
In nes_upost_send, the sq_head value is picked up before the qp lock
had been acquired.  This would allow two threads to fill in the same
location in the SQ.

The same problem exists in nes_upost_recv for the RQ.

This fix makes sure the head value is picked up after the lock is taken
in both routines.

Signed-off-by: Don Wood <donald.e.wood@intel.com>