Perfquery can be too noisy.
When perfquery is run against fabrics that do not support PortXmitWait, it emits this warning for every port:

ibwarn: [23225] dump_perfcounters: PortXmitWait not indicated so ignore this counter

When running ibcheckerrors on a large fabric, this leads to a flood of warnings.

The proposed patch reduces the warning to a verbose message.

Signed-off-by: Michael Heinz <michaelheinz@qlogic.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>