management: adding mad_dump_fields to libibmad
Sasha asked that I re-submit the patches for perfquery in a slightly different format. This is the first of 3 patches.

This patch adds a function to libibmad that allows the caller to dump a configurable range of MAD attributes. Basically, this provides an external interface to the internal function _dump_fields.

Signed Off: Michael Heinz

---------------- snip --------------

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>