ib/mgmt: fix build under windows stack
Fix warnings in windows build.

These are typecasts in several places, with a couple additional changes
needed to libibnetdisc.  Windows uses ISO definitions instead of
POSIX definitions, so calls like _read, _write, etc. are used on
Windows instead of read, write.  The result is that internal functions
are renamed with an ibnd prefix.

In order to handle ibdebug exported from libibmad, Windows need to mark
the variable as a 'dllexport' from libibmad, but as 'dllimport' for users.
Since libibnetdisc wants to 'dllimport' ibdebug from libibmad, while
'dllexporting its own functions, it requires a separate 'MAD_EXPORT' like
definition to allow this.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>