infiniband-diags/ibnetdiscover: Fix handling of CA ports in recv_port_info
When multiple ports on the same CA are connected to the same subnet and
an ibnetdiscover is initiated from one of them, the discovery continues
past the other CA port and an error occurs.

The error is:
src/query_smp.c:188; umad (DR path slid 0; dlid 0; 0,1,20,2 Attr0x11:0) bad status 110; Connection timed out

Fix this by saving the initiating port number in the fabric structure
and using that rather than local_port for the comparison in
recv_port_info.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>