librdmacm: only allocate qp in rdma_create_ep if qp_attr provided
The comments and documentation for rdma_create_ep indicate that
it will only allocate a QP if initial QP attributes are provided.
However, the code always attempts to create a QP off an associated
active rdma_cm_id endpoint.

By _not_ allocating the QP, this allows a user to first determine
what RDMA device a rdma_cm_id was associated with.  The user can
then create a QP that references an existing CQ, SRQ, or PD.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>