ibacm: automatically generate addresses if missing acm_addr.cfg
If the acm_addr.cfg file is missing, automatically attempt to
generate the file.  If no addresses can be assigned to any
endpoints because of a missing file, log an error and exit
the service.  This avoids running a service that is basically
useless.

This change allows the ib_acm service to execute using defaults
without requireing that a user first run 'ib_acme -A -O' or
otherwise creating the acm configuration files.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>