ib_acme: add an 'unspecified' address format
Support an 'unspecified' address format.  If no address
format is given, ib_acme will use getaddrinfo to try to
resolve the destination address.

This simplifies the user interface, plus allows the user
to specify a hostname for input, but allow getaddrinfo
to resolve the hostname into an IP address before contacting
the ACM service.  The latter can provide a benefit of
being able to specify hostnames on a large cluster, but
allow ib_acme to populate ib_acm caches with IP address
data instead.  This benefits MPI applications which may
exchange IP addresses out of band for communications, but
rely on hostnames for job distribution.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>